pyproject.toml
[project]
name = "rye_template"
version = "0.0.1"
description = "Add a short description here"
authors = []
dependencies = [
    "pydantic>=2.6.4,<3.0.0",
    "pydantic-settings>=2.2.1,<3.0.0",
    "pyyaml>=6.0.1",
]
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.9.0,<2.0.0",
    "ruff>=0.3.4",
    "pytest>=8.1.1",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.2",
]

[tool.rye.scripts]
format = { chain = ["ruff check --fix src"] }
lint = { chain = ["ruff check src", "mypy src"] }
test = { chain = ["python3 -m pytest tests"] }

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false
disallow_untyped_calls = false

[tool.ruff]
src = ["src", "tests"]
exclude = [".venv"]
lint.select = [
    "F", # pyflakes
    "E", # pycodestyle
    "W", # pycodestyle warnings
    "I", # isort
    "D", # pydocstyle
]
lint.ignore = []
lint.extend-ignore = [
    "B008", # Dependsで使用するため
    "D100", # missing-module-docstring
    "D103", # missing-docstring
    "D104", # missing-class-docstring
    "D105", # undocumented-magic-method
    "D106", # missing-function-docstring
    "D107", # undocumented-public-init
    "D205", # blank-line-after-summary
    "D415", # ends-in-punctuation
]
# 1行の最大文字数
line-length = 88

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
# 個別設定
# __init__.pyは未使用インポートを許容
"__init__.py" = ["F401"]
